FIXED: Manual integration of files GUI.h and Gui.h (these are considered the same file on Windows). In the original code, these files seem to contain almost the same content, bit with small variations.
