+++ Simple_window.h
@@ -1,5 +1,5 @@
 
-#include "GUI.h"	// for Simple_window only (doesn't really belong in Window.h)
+#include "GUI.hpp"	// for Simple_window only (doesn't really belong in Window.h)
 
 using namespace Graph_lib;
